requirements.txt
loguru
sqlalchemy
pyyaml
requests
pytest
jinja2
